arm/requirements.txt
requests
packaging
semver
python-dotenv